move login event to global events folder
